Make object v-table internal
