+++ src/api.h
@@ -311,26 +311,12 @@ typedef void (ag_object_list_iterator_mutable)(ag_object_t **node, void *opt);
 #define ag_object_list_smart ag_object_smart_t
 
 
-                                      /* initialises object v-table [AgDM:??] */
-extern void ag_object_vtable_init(size_t len);
+                                       /* initialises object system [AgDM:??] */
+extern void ag_object_init(size_t len);
 
 
-                                       /* shuts down object v-table [AgDM:??] */
-extern void ag_object_vtable_exit(void);
-
-
-                         /* checks if methods for object type exist [AgDM:??] */
-extern ag_pure bool ag_object_vtable_exists(unsigned type);
-
-
-                                     /* gets methods of object type [AgDM:??] */
-extern ag_pure ag_hot const struct ag_object_method *ag_object_vtable_get(
-        unsigned type);
-
-
-                                     /* sets methods of object type [AgDM:??] */
-extern void ag_object_vtable_set(unsigned type, 
-        const struct ag_object_method *meth);
+                                        /* shuts down object system [AgDM:??] */
+extern void ag_object_exit(void);
 
 
                                                 /* registers object [AgDM:??] */
